Issue #178: Fix Sonar - split test class and fix constant declarations
ETP-3623
